fix: formato fecha en historial de pagos actualizada

+++ src/features/payments/Payments.jsx
@@ -173,7 +173,7 @@ export const Payments = () => {
                                                 </span>
                                             </td>
                                             <td className='py-3 px-4 text-muted text-xs'>
-                                                {formatDate(tx.created_at,{ year: 'numeric', month: 'short', day: 'numeric' })}
+                                                {formatDate(tx.created_at,{ year: 'numeric', month: 'numeric', day: 'numeric' })}
                                             </td>
                                         </tr>
                                     ))}
